いい感じ

+++ DirectXGame/engine/effect/Particle/ParticleManager.h
@@ -120,7 +120,7 @@ public:
 
 
 
-		float count;
+		int count;
 	};
 
 	struct Particle
